7.2.1 Some refactoring. Work completed in previous step
